requirements.txt
asyncio==3.4.3
paho-mqtt==1.5.1
websockets==8.1
